GRE tunnels should include local_ip.
This addresses the issue of having one or more multi-homed host.

Because you may only specify one local IP for a quantum agent, and
other hosts could be routed across different subnets (even for small
installations, for various reasons), it's important to also specify
the used local_ip when setting up the GRE tunnels. As long as the
address is routable on both ends, this will work. If the local_ip is
not specified, then traffic will mysteriously be dropped on one end
where the IP does not match the expected IP in the GRE tunnel.

Bug: https://bugs.launchpad.net/quantum/+bug/1184696

Change-Id: I5d612bf9e4f6652af8b155cdd1748d73ac4539ff
Signed-off-by: Adin Scannell <adin@scannell.ca>