Added config option for toggling hugepage allocation
If hugepages are allocated before ovs-dpdk is run, e.g. with a container,
then the service should not attempt to allocate hugepages when started,
nor should it attempt to de-allocate them when stopping.

To prevent hugepages from being allocated/deallocated by the service,
the option OVS_ALLOCATE_HUGEPAGES=False should be set.

Change-Id: I51421a1b72c1a14a199424cd47f13d6f618061ab