set ERROR_ON_CLONE to False for the plugin
ERROR_ON_CLONE is set to True in devstack-gate script
preventing this plugin to clone source code for openvswitch
and dpdk. ERROR_ON_CLONE in devstack-gate is mainly used
to protect openstack* repositories from being cloned instead of
being taken from it's own filesystem prepared for devstack-gate
uses. Such protection is not necessary for this plugin.

Change-Id: Ic9ef303958d4e063329bf40d0de2c4d190eb098b
Closes-Bug: #1573826