New Nova Scheduler filter: AggregateInstanceTypeFilter
This blueprint introduces a new filter to the nova filter scheduler.
This new filter will deprecate the current AggregateInstanceExtraSpecsFilter
filter. This new filter, AggregateInstanceTypeFilter, will be backwards
compatible with the deprecated one and will introduce the use of sentinel
values to add more flexibility during the process of creating filtering rules.

blueprint aggregate-instance-type-filter

Change-Id: Ieb1f5adca6c61a16ae21d117fd728299d0efb2d9